Use redis client builder
