+++ vcxagency-node/src/service/notifications/service-new-messages.js
@@ -40,43 +40,6 @@ module.exports.createServiceNewMessages = function createServiceNewMessages (red
     redisClientRw.quit()
   }
 
-  redisClientRw.on('error', function (err) {
-    logger.error(`Redis rw-client encountered error: ${err}`)
-  })
-  redisClientSubscriber.on('error', function (err) {
-    logger.error(`Redis subscription-client encountered error: ${err}`)
-  })
-
-  redisClientRw.on('end', () => {
-    console.log('Redis rw-client disconnected')
-  })
-  redisClientSubscriber.on('end', () => {
-    console.log('Redis subscription-client disconnected')
-  })
-
-  redisClientRw.on('reconnecting', () => {
-    console.log('Redis rw-client reconnecting')
-  })
-
-  redisClientSubscriber.on('reconnecting', () => {
-    console.log('Redis subscription-client reconnecting')
-  })
-
-  redisClientRw.on('connect', function () {
-    logger.info('Redis rw-client connected.')
-  })
-  redisClientSubscriber.on('connect', function () {
-    logger.info('Redis subscription-client connected.')
-  })
-
-  redisClientSubscriber.on('subscribe', function (channel, _count) {
-    logger.info(`Subscribed on channel ${channel}.`)
-  })
-
-  redisClientSubscriber.on('unsubscribe', function (channel, _count) {
-    logger.info(`Unsubscribed on channel ${channel}.`)
-  })
-
   redisClientSubscriber.on('message', async function (channel, message) {
     try {
       if (isKeyspaceSetNotification(channel, message)) {
